First Injection Bug Fixed
Bug where the channels were not opened for the first injection after the
switch board was turned on. Cause was incorrect timing of the reset and
sync pins, so the switches were not in a receptive state until the
second injection.

Also, added delay to start of injection as CS takes ~50ms to actually
start injecting. Trigger channels now correspond to when the current
*really* starts, so no changes in processing needed